requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
objaverse
google-generativeai
langchain
ratelimit 
streamlit 
plotly 
pandas 
numpy 
scikit-learn